% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{augment_with_monday_id}
\alias{augment_with_monday_id}
\title{Augment rin_service_areas with monday_id column (one-time migration)}
\usage{
augment_with_monday_id(params, old_rin_service_areas)
}
\arguments{
\item{params}{An object containing the rin_community_alias_to_item_id mapping}

\item{old_rin_service_areas}{data.frame from cori.data.rin::rin_service_areas}
}
\value{
data.frame with monday_id prepended as first column
}
\description{
Takes the existing package data and adds a monday_id column by looking up
each rin_community in the rin_community_alias_to_item_id mapping from params.
}
