% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_api.R
\name{fetch_monday_board}
\alias{fetch_monday_board}
\title{Fetch all data from a Monday.com board}
\usage{
fetch_monday_board(board_id, group_title)
}
\arguments{
\item{board_id}{Monday board ID}

\item{group_title}{Title of the group to fetch items from (e.g., "Current")}
}
\value{
data.frame with all columns from the board
}
\description{
Fetch all data from a Monday.com board
}
