% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{geocode_rin_map_data}
\alias{geocode_rin_map_data}
\title{Geocode RIN map data using OpenStreetMap}
\usage{
geocode_rin_map_data(communities_data, rin_service_areas)
}
\arguments{
\item{communities_data}{Tibble from load_comms_communities()}

\item{rin_service_areas}{Tibble from load_rin_service_areas() - used to filter to valid communities}
}
\value{
Tibble with lat and long columns added
}
\description{
Geocode RIN map data using OpenStreetMap
}
