% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_comms_communities}
\alias{load_comms_communities}
\title{Load RIN communities from Comms Google Sheet}
\usage{
load_comms_communities(params = cori.utils::get_params("global"))
}
\arguments{
\item{params}{Parameters containing comms_sheet_id and comms_sheet_name}
}
\value{
Transformed tibble with rin_community, community_name, geocode_column, url
}
\description{
Load RIN communities from Comms Google Sheet
}
