% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_rin_service_areas}
\alias{load_rin_service_areas}
\title{A function to generate RIN service areas (county level) from an XLSX extract from Monday (see params.yml)}
\usage{
load_rin_service_areas(params, old_rin_service_areas)
}
\arguments{
\item{params}{An object containing values for the $current_year and $monday_network_communities_file_name parameters}

\item{old_rin_service_areas}{data.frame of previous RIN service areas build}
}
\value{
data.frame of all counties associated with each RIN community
}
\description{
A function to generate RIN service areas (county level) from an XLSX extract from Monday (see params.yml)
}
