% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{lookup_monday_id}
\alias{lookup_monday_id}
\title{Look up monday_id for each name using the rin_community_alias_to_item_id mapping}
\usage{
lookup_monday_id(names_vec, alias_list)
}
\arguments{
\item{names_vec}{character vector of rin_community names to look up}

\item{alias_list}{named list from params$rin_community_alias_to_item_id}
}
\value{
character vector of monday_id values, same length as names_vec
}
\description{
Look up monday_id for each name using the rin_community_alias_to_item_id mapping
}
